process

+++ models/parsing.go
@@ -1,14 +1,14 @@
 package models
 
 type ParseParams struct {
-	Limit int32  `json:"limit"`
+	Limit int  `json:"limit"`
 	Since string `json:"since"`
 	Desc  bool   `json:"desc"`
 }
 
 type ParseParamsThread struct {
-	Limit int32  `json:"limit"`
-	Since int64  `json:"since"`
+	Limit int  `json:"limit"`
+	Since int  `json:"since"`
 	Sort  string `lson:"since"`
 	Desc  bool   `json:"desc"`
 }
